Add support for strong consistency on GetWorkflowHistory and DescribeWorkflowExecution (#1445)
What changed?

Adds support to make requests with a QueryConsistencyLevel to:

GetWorkflowHistory
DescribeWorkflowExecution
Using STRONG as the QueryConsistencyLevel will direct cadence to retrieve information from the "active" cluster, regardless of where the request is received (active or passive). Note that this will increase the latency of requests that are received in the passive cluster.

The existing methods have been kept for compatibility and can continue to be used.